fix: resolve all compilation errors and launch desktop app successfully
- Fixed final 2 Send trait errors by restructuring database transactions
- Updated Dioxus syntax for version 0.5 compatibility
- Moved CSS styles to const to avoid rsx! macro parsing issues
- Fixed all signal access patterns (read/write methods)
- Desktop app now compiles and runs successfully

✅ All compilation errors resolved (320 → 0)
✅ Desktop app launches with VS Code-style interface

+++ src/analytics/dashboard.rs
@@ -880,7 +880,7 @@ impl WidgetRenderer for LineChartRenderer {
                     let min_val = points.iter().map(|p| p.value).fold(f64::INFINITY, f64::min);
                     
                     // Create simple sparkline
-                    let sparkline = create_sparkline(points, width - 4);
+                    let sparkline = create_sparkline(points, (width - 4) as usize);
                     output.push_str(&format!("│ {} │\n", sparkline));
                     output.push_str(&format!("│ Min: {:.1} Max: {:.1}{:>width$} │\n", 
                         min_val, max_val, "",
